Add UI showing how much of the map has been planted
Added a bar in the UI which indicates to the player how much of the map is covered with empty Ground, Water, Plants and fully grown/Bloomed Plants.

Updated the related map statistics functions to be more performant by changing single tiles at a time instead of recalculating the entire map every time a change is made.