Add Quest System and Add Choices UI to Dialogue
Add the beginnings of a Quest System, with quests that can have multiple quest goals.
Quest goals are updated via the GameEvents system.

Update the Dialogue handling so that choices in the dialogue are displayed as buttons.

The player can no longer move etc. while in a dialogue.

Moved plants scripts to a new folder to improve project structure.